Implemented basic data uploading.

+++ PD/Controllers/UploadController.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using Hangfire;
 using Microsoft.AspNetCore.DataProtection;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -50,35 +51,69 @@ namespace PD.Controllers
                         await formFile.CopyToAsync(stream);
                     }
 
-                    ////BackgroundService srv = new BackgroundService(_context);
-                    ////string jobKey = "Faculty Data import";
-                    ////string jobId = srv.Enque(() => 
-                    ////    new ImportService(new BackgroundService(_configuration).CreateDbContext())
-                    ////        .InjestFacultySalaryAdjustmentData(tmpFile, worksheetName, new DateTime(2015, 7, 1).Date,
-                    ////            new DateTime(2016, 06, 30).Date, new DateTime(2017, 06, 30).Date, true),
-                    ////    jobKey);
+                    string jobKey = "Faculty Data import";
+                    if (_importService.IsJobActive(jobKey))
+                    {
+                        System.IO.File.Delete(tmpFile);
+                        ViewBag.Message = "Previous job is still running. No new job added.";
+                        ViewBag.MessageClass = "alert-danger";
+                    }
+                    else
+                    {
+                        var jobId = BackgroundJob.Enqueue<ImportService>(srv => srv.InjestFacultySalaryAdjustmentData(
+                            tmpFile,
+                            worksheetName,
+                            new DateTime(2015, 7, 1).Date,
+                            new DateTime(2016, 06, 30).Date,
+                            new DateTime(2017, 06, 30).Date,
+                            true
+                            ));
 
-                    ////if(string.IsNullOrEmpty(jobId))
-                    ////{
-                    ////    ViewBag.Message = "Previous job is still running. No new job added.";
-                    ////    ViewBag.MessageClass = "alert-danger";
-                    ////    System.IO.File.Delete(tmpFile);
-                    ////}
-                    ////else
-                    ////{
-                    ////    ViewBag.Message = string.Format("New job Added. Job ID: {0}", jobId);
-                    ////    ViewBag.MessageClass = "alert-info";
-                    ////}
+                        //Saving the job ID
+                        var job = _context.HangFireJobs.Where(j => j.JobKey == jobKey).FirstOrDefault();
+                        if (job == null)
+                        {
+                            job = new Models.HangFireJob() { JobId = jobId, JobKey = jobKey };
+                            _context.HangFireJobs.Add(job);
+                        }
+                        else
+                            job.JobId = jobId;
+                        _context.SaveChanges();
+
+                        ViewBag.Message = string.Format("New job Added. Job ID: {0}", jobId);
+                        ViewBag.MessageClass = "alert-info";
+                    }
+
+                    ViewBag.FilePath = tmpFile;
+
+                    ////////BackgroundService srv = new BackgroundService(_context);
+                    ////////string jobId = srv.Enque(() =>
+                    ////////    new ImportService(new BackgroundService(_configuration).CreateDbContext())
+                    ////////        .InjestFacultySalaryAdjustmentData(tmpFile, worksheetName, new DateTime(2015, 7, 1).Date,
+                    ////////            new DateTime(2016, 06, 30).Date, new DateTime(2017, 06, 30).Date, true),
+                    ////////    jobKey);
+
+                    ////////if (string.IsNullOrEmpty(jobId))
+                    ////////{
+                    ////////    ViewBag.Message = "Previous job is still running. No new job added.";
+                    ////////    ViewBag.MessageClass = "alert-danger";
+                    ////////    System.IO.File.Delete(tmpFile);
+                    ////////}
+                    ////////else
+                    ////////{
+                    ////////    ViewBag.Message = string.Format("New job Added. Job ID: {0}", jobId);
+                    ////////    ViewBag.MessageClass = "alert-info";
+                    ////////}
 
                     ViewBag.FilePath = tmpFile;
 
 
-                    //processing the uploaded file
-                    _importService.InjestFacultySalaryAdjustmentData(tmpFile, worksheetName, new DateTime(2015, 7, 1).Date,
-                        new DateTime(2016, 06, 30).Date, new DateTime(2017, 06, 30).Date);
+                    //////processing the uploaded file
+                    ////_importService.InjestFacultySalaryAdjustmentData(tmpFile, worksheetName, new DateTime(2015, 7, 1).Date,
+                    ////    new DateTime(2016, 06, 30).Date, new DateTime(2017, 06, 30).Date);
 
-                    //Deleting the temporary file
-                    System.IO.File.Delete(tmpFile);
+                    //////Deleting the temporary file
+                    ////System.IO.File.Delete(tmpFile);
                 }
             }
